artifact/04-analysis-scripts/pre-scan/requirements.txt
# Python dependencies for the pre-scan scripts
dnspython
tldextract
requests
pyasn
